// client/src/draft.tsx
export type Draft = {
    // stores each possible pick of the draft as an element
    pick: string[]
    // stores each team name of the draft as an element
    team: string[]
    // stores the id of the draft 
    id: number  | undefined
    // stores the number of rounds in the draft
   rounds: number  | undefined
   // stores the drafter's name
   drafter: string
  }

 

 
  